--- a/constantine.nimble
+++ b/constantine.nimble
@@ -403,8 +403,8 @@ const testDesc: seq[tuple[path: string, useGMP: bool]] = @[
   # ("tests/math_fields/t_finite_fields.nim", false),
   # ("tests/math_fields/t_finite_fields_conditional_arithmetic.nim", false),
   ("tests/math_fields/t_finite_fields_mulsquare.nim", false),
-  # ("tests/math_fields/t_finite_fields_sqrt.nim", false),
-  # ("tests/math_fields/t_finite_fields_powinv.nim", false),
+  ("tests/math_fields/t_finite_fields_sqrt.nim", false),
+  ("tests/math_fields/t_finite_fields_powinv.nim", false),
   ("tests/math_fields/t_finite_fields_vs_gmp.nim", true),
   # ("tests/math_fields/t_fp_cubic_root.nim", false),