--- a/constantine/math/arithmetic/assembly/limbs_asm_mul_mont_x86.nim
+++ b/constantine/math/arithmetic/assembly/limbs_asm_mul_mont_x86.nim
@@ -33,7 +33,7 @@ static: doAssert UseASM_X86_64
 macro mulMont_CIOS_sparebit_gen[N: static int](
         r_PIR: var Limbs[N], a_PIR, b_PIR,
         M_MEM: Limbs[N], m0ninv_REG: BaseType,
-        skipFinalSub: static bool): untyped =
+        skipFinalReduction: static bool): untyped =
   ## Generate an optimized Montgomery Multiplication kernel
   ## using the CIOS method
   ##
@@ -161,21 +161,21 @@ macro mulMont_CIOS_sparebit_gen[N: static int](
   ctx.mov rax, r # move r away from scratchspace that will be used for final substraction
   let r2 = rax.asArrayAddr(r_PIR, len = N, memIndirect = memWrite)
 
-  if skipFinalSub:
+  if skipFinalReduction:
     for i in 0 ..< N:
       ctx.mov r2[i], t[i]
   else:
     ctx.finalSubNoOverflowImpl(r2, t, M, scratch)
   result.add ctx.generate()
 
-func mulMont_CIOS_sparebit_asm*(r: var Limbs, a, b, M: Limbs, m0ninv: BaseType, skipFinalSub: static bool = false) =
+func mulMont_CIOS_sparebit_asm*(r: var Limbs, a, b, M: Limbs, m0ninv: BaseType, skipFinalReduction: static bool = false) =
   ## Constant-time Montgomery multiplication
-  ## If "skipFinalSub" is set
+  ## If "skipFinalReduction" is set
   ## the result is in the range [0, 2M)
   ## otherwise the result is in the range [0, M)
   ##
   ## This procedure can only be called if the modulus doesn't use the full bitwidth of its underlying representation
-  r.mulMont_CIOS_sparebit_gen(a, b, M, m0ninv, skipFinalSub)
+  r.mulMont_CIOS_sparebit_gen(a, b, M, m0ninv, skipFinalReduction)
 
 # Montgomery Squaring
 # ------------------------------------------------------------
@@ -184,11 +184,11 @@ func squareMont_CIOS_asm*[N](
        r: var Limbs[N],
        a, M: Limbs[N],
        m0ninv: BaseType,
-       spareBits: static int, skipFinalSub: static bool) =
+       spareBits: static int, skipFinalReduction: static bool) =
   ## Constant-time modular squaring
   var r2x {.noInit.}: Limbs[2*N]
   square_asm(r2x, a)
-  r.redcMont_asm(r2x, M, m0ninv, spareBits, skipFinalSub)
+  r.redcMont_asm(r2x, M, m0ninv, spareBits, skipFinalReduction)
 
 # Montgomery Sum of Products
 # ------------------------------------------------------------
@@ -196,7 +196,7 @@ func squareMont_CIOS_asm*[N](
 macro sumprodMont_CIOS_spare2bits_gen[N, K: static int](
         r_PIR: var Limbs[N], a_PIR, b_PIR: array[K, Limbs[N]],
         M_MEM: Limbs[N], m0ninv_REG: BaseType,
-        skipFinalSub: static bool): untyped =
+        skipFinalReduction: static bool): untyped =
   ## Generate an optimized Montgomery merged sum of products ⅀aᵢ.bᵢ kernel
   ## using the CIOS method
   ##
@@ -353,7 +353,7 @@ macro sumprodMont_CIOS_spare2bits_gen[N, K: static int](
   ctx.mov rax, r # move r away from scratchspace that will be used for final substraction
   let r2 = rax.asArrayAddr(r_PIR, len = N, memIndirect = memWrite)
 
-  if skipFinalSub:
+  if skipFinalReduction:
     ctx.comment "  Copy result"
     for i in 0 ..< N:
       ctx.mov r2[i], t[i]
@@ -367,11 +367,11 @@ macro sumprodMont_CIOS_spare2bits_gen[N, K: static int](
 func sumprodMont_CIOS_spare2bits_asm*[N, K: static int](
         r: var Limbs[N], a, b: array[K, Limbs[N]],
         M: Limbs[N], m0ninv: BaseType,
-        skipFinalSub: static bool) =
+        skipFinalReduction: static bool) =
   ## Sum of products ⅀aᵢ.bᵢ in the Montgomery domain
-  ## If "skipFinalSub" is set
+  ## If "skipFinalReduction" is set
   ## the result is in the range [0, 2M)
   ## otherwise the result is in the range [0, M)
   ##
   ## This procedure can only be called if the modulus doesn't use the full bitwidth of its underlying representation
-  r.sumprodMont_CIOS_spare2bits_gen(a, b, M, m0ninv, skipFinalSub)
+  r.sumprodMont_CIOS_spare2bits_gen(a, b, M, m0ninv, skipFinalReduction)