--- a/constantine/math/arithmetic/assembly/limbs_asm_mul_mont_x86_adx_bmi2.nim
+++ b/constantine/math/arithmetic/assembly/limbs_asm_mul_mont_x86_adx_bmi2.nim
@@ -62,6 +62,8 @@ proc mulx_by_word(
 
   # Steady state
   for j in 1 ..< N-1:
+    # TODO: should we alternate lo with another register?
+    #       to deal with false dependencies?
     ctx.mulx t[j+1], lo, a[j], rdx
     if j == 1:
       ctx.add t[j], lo