--- a/constantine/math/arithmetic/assembly/limbs_asm_mul_x86_adx_bmi2.nim
+++ b/constantine/math/arithmetic/assembly/limbs_asm_mul_x86_adx_bmi2.nim
@@ -52,6 +52,8 @@ proc mulx_by_word(
   ctx.mov r0, rax
 
   # Steady state
+  # TODO: should we alternate rax with another register?
+  #       to deal with false dependencies?
   for j in 1 ..< N:
     ctx.mulx t[j], rax, a[j], rdx
     if j == 1: