… PLANNING.md
… benchmarks/bench_elliptic_parallel_template.nim
… benchmarks/bench_elliptic_template.nim
… benchmarks/bench_fields_template.nim
… benchmarks/bench_fp.nim
… constantine/math/arithmetic/assembly/limbs_asm_redc_mont_x86_adx_bmi2.nim
… constantine/math/arithmetic/bigints_crandall.nim
… constantine/math/arithmetic/bigints_montgomery.nim
… constantine/math/arithmetic/finite_fields.nim
… constantine/math/arithmetic/limbs_crandall.nim
… constantine/math/arithmetic/limbs_montgomery.nim
… constantine/math/elliptic/ec_multi_scalar_mul_scheduler.nim
… constantine/math/elliptic/ec_shortweierstrass_batch_ops.nim
… constantine/math/elliptic/ec_shortweierstrass_jacobian.nim
… constantine/math/elliptic/ec_shortweierstrass_jacobian_extended.nim
… constantine/math/elliptic/ec_shortweierstrass_projective.nim
… constantine/math/elliptic/ec_twistededwards_batch_ops.nim
… constantine/math/extension_fields/towers.nim
… constantine/math_arbitrary_precision/arithmetic/limbs_montgomery.nim
… constantine/named/config_fields_and_curves.nim
… constantine/named/deriv/derive_constants.nim
… constantine/named/deriv/parser_fields.nim
… constantine/named/deriv/precompute.nim
… constantine/named/properties_fields.nim
… constantine/platforms/x86/macro_assembler_x86_att.nim
… constantine/platforms/x86/macro_assembler_x86_intel.nim
… metering/eip2537.md
… tests/math_elliptic_curves/t_ec_shortw_jac_g1_add_double.nim
… tests/math_elliptic_curves/t_ec_shortw_jac_g1_mixed_add.nim
… tests/math_elliptic_curves/t_ec_shortw_jac_g1_mul_sanity.nim