This closes #11 for primes of form 2ᵐ-c (Crandall primes / pseudo-Mersenne primes), such as the one used for Curve25519 and secp256kq (Ethereum/ Bitcoin).

## Bench Fp vs Constantine master

### Previous

![image](https://github.com/user-attachments/assets/17801587-6dc6-435e-80a8-1ccf06312f62)
![image](https://github.com/user-attachments/assets/9f9da087-f86e-4a6b-b1d7-7a1ff3446899)

### Current

![image](https://github.com/user-attachments/assets/b316ee76-03fe-4272-af27-42b2250e15fe)
![image](https://github.com/user-attachments/assets/f651423f-ba6f-4f9b-b005-9f3c2bbd9653)

### Analysis

- Fp[Edwards25519] mul 1.27x improvement
- Fp[Edwards25519] square 1.43x improvement
- Fp[Secp256k1] mul 1.94x improvement
- Fp[Secp256k1] square 1.28x improvement

## Bench EC vs Constantine master

### Previous

![image](https://github.com/user-attachments/assets/96884ddd-b302-42be-b5ab-1f392a9bd12c)

### Current

![image](https://github.com/user-attachments/assets/ddef560b-bab4-42d0-bbb5-f47a557e9214)

### Analysis

- EC add projective constant-time improved by 1.36x
- EC add jacobian constant-time improved by 1.34x
- EC add projective vartime improved by 1.24x
- EC add jacobian vartime improved by 1.37x
- EC dbl projective constant-time improved by 1.31x
- EC dbl jacobian constant-time improved by 1.06x

## Bench vs bitcoin/secp256k1

![image](https://github.com/user-attachments/assets/a8258d60-8fad-49de-9114-2b36a25666c4)
- field_sqr 12.4ns vs 8ns -> 1.55x
- field_mul 15.8ns vs 10ns -> 1.58x
- field_inv_ct 1410ns vs 1203ns -> 1.17x
- field_inv_vt 820ns vs 848ns -> 0.97x
- EC add jacobian var 247ns vs 97ns -> 2.55x
- EC dbl jacobian var 97.8ns vs 145 -> 0.67x
- EC mixed add ct 189ns vs 225ns -> 0.84x
- EC mixed add var 173ns vs 98ns -> 1.77x

![image](https://github.com/user-attachments/assets/1e8d0eb9-b707-43f1-8c65-b8200433f8c3)

- EC scalar-mul ct  28100ns vs  40196 ns -> 0.70x

## Analysis

The fact that field operations are 1.5x faster BUT the elliptic curve operations are sometimes slower is suspicious. We probably need to check the EC formulae

## TODO

- fix windows
- bound checks for lazy reduce and lazy reduced field exponentiation for 256-bit as eprint/iacr 2018/985
  indicates in Theorem 4 that their partial reduction may grow by 1 bit if 256-bit.
- optimize EC impl to avoid if/else check for ADX and limit input/output movement
- optimized mixed add